# api/openapi.yaml
openapi: 3.1.0
info:
  title: ATOS REST API
  version: 0.2.0
  description: |
    REST projection of the ATOS Agent Internet gateway. Managed, Verified and
    Native share one API. `requested_trust_mode=auto` is accepted before Quote
    creation; every Quote freezes one concrete `trust_mode`, which propagates
    through Job, Escrow, Execution Receipt and settlement.

    This implementation currently exposes the complete Managed flow. Verified
    and Native contracts are modeled but fail closed until their real TOS
    integrations are available.
servers:
  - url: https://api.atos.im
    description: Production reference gateway
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /activate:
    get:
      tags: [auth]
      summary: Device Authorization verification page
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
      responses:
        '200':
          description: Human-readable authorization status
          content:
            text/html:
              schema: {type: string}
    post:
      tags: [auth]
      summary: Record a browser consent decision behind the trusted ATOS login proxy
      security: []
      parameters:
        - name: X-ATOS-Approval-Token
          in: header
          required: true
          schema: {type: string}
        - name: X-ATOS-Principal-ID
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: false
              required: [user_code, decision, csrf_token]
              properties:
                user_code: {type: string}
                decision: {type: string, enum: [approve, deny]}
                csrf_token: {type: string, minLength: 64, maxLength: 64}
      responses:
        '303': {description: Decision recorded and redirected to the activation status page}
        default: {$ref: '#/components/responses/Error'}
  /confirm:
    get:
      tags: [confirmations]
      summary: Human-readable spend-confirmation page
      security: []
      parameters:
        - {name: code, in: query, schema: {type: string}}
      responses:
        '200':
          description: Confirmation instructions
          content:
            text/html:
              schema: {type: string}
  /v1/auth/device:
    post:
      tags: [auth]
      summary: Start scoped Device Authorization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAuthorizationRequest'
      responses:
        '200':
          description: Device grant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceAuthorizationResponse'
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/device/token:
    post:
      tags: [auth]
      summary: Exchange an authorized device code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [device_code]
              properties:
                device_code: {type: string, minLength: 1}
      responses:
        '200':
          description: Scoped access and refresh tokens
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        '400':
          description: OAuth Device Authorization status (`authorization_pending`, `slow_down`, `access_denied`, or `expired_token`)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OAuthError'}
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/device/decision:
    post:
      tags: [auth]
      summary: Trusted atos.im consent surface approves or denies a device code
      security: []
      parameters:
        - name: X-ATOS-Approval-Token
          in: header
          required: true
          schema: {type: string}
        - name: X-ATOS-Principal-ID
          in: header
          required: true
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [user_code, decision]
              properties:
                user_code: {type: string}
                decision: {type: string, enum: [approve, deny]}
      responses:
        '200': {description: Device decision recorded}
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/token/refresh:
    post:
      tags: [auth]
      summary: Rotate a refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [refresh_token]
              properties:
                refresh_token: {type: string, minLength: 1}
      responses:
        '200':
          description: Rotated token pair
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/revoke:
    post:
      tags: [auth]
      summary: Revoke the current access token
      responses:
        '204': {description: Revoked}
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/devices:
    get:
      tags: [auth]
      summary: List the caller's authorized devices
      responses:
        '200':
          description: Device inventory
          content:
            application/json:
              schema:
                type: object
                required: [devices]
                properties:
                  devices: {type: array, items: {$ref: '#/components/schemas/AuthorizedDevice'}}
        default: {$ref: '#/components/responses/Error'}
  /v1/auth/devices/{id}:
    delete:
      tags: [auth]
      summary: Revoke one authorized device and all of its credentials
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '204': {description: Device revoked}
        default: {$ref: '#/components/responses/Error'}
  /v1/capabilities:
    get:
      tags: [capabilities]
      summary: Search active Capabilities
      description: Requires `capabilities:read`.
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: max_price, in: query, schema: {type: string}}
        - {name: max_price_currency, in: query, schema: {type: string, default: USD}}
        - {name: delivery_mode, in: query, schema: {type: array, items: {type: string, enum: [instant, async, interactive]}}, explode: true}
        - {name: min_trust_score, in: query, schema: {type: number, minimum: 0, maximum: 1}}
        - {name: max_latency_ms, in: query, schema: {type: integer, minimum: 0}}
        - {name: requested_trust_mode, in: query, schema: {$ref: '#/components/schemas/RequestedTrustMode'}}
        - {name: network_verifiable_receipt, in: query, schema: {type: boolean}}
        - {name: tos_settlement, in: query, schema: {type: boolean}}
        - {name: portable_proof_of_service, in: query, schema: {type: boolean}}
      responses:
        '200':
          description: Ranked Capability projection
          content:
            application/json:
              schema:
                type: object
                required: [capabilities]
                properties:
                  capabilities:
                    type: array
                    items: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [capabilities]
      summary: Register a provider Capability
      description: Requires `capabilities:write`. Requested stronger modes remain pending until certified.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterCapabilityRequest'}
      responses:
        '201':
          description: Capability registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
  /v1/capabilities/{id}:
    parameters:
      - {$ref: '#/components/parameters/CapabilityID'}
    get:
      tags: [capabilities]
      summary: Get one Capability
      description: Requires `capabilities:read`.
      responses:
        '200':
          description: Capability
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
    patch:
      tags: [capabilities]
      summary: Update mutable Capability fields
      description: Requires `capabilities:write` and target ownership. Derived active mode fields are immutable through this route.
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, additionalProperties: true}
      responses:
        '200':
          description: Updated Capability
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
  /v1/capabilities/{id}/pause:
    parameters:
      - {$ref: '#/components/parameters/CapabilityID'}
    post:
      tags: [capabilities]
      summary: Pause an owned Capability
      responses:
        '200':
          description: Paused Capability
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
  /v1/capabilities/{id}/resume:
    parameters:
      - {$ref: '#/components/parameters/CapabilityID'}
    post:
      tags: [capabilities]
      summary: Resume an owned Capability
      responses:
        '200':
          description: Active Capability
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Capability'}
        default: {$ref: '#/components/responses/Error'}
  /v1/quotes:
    post:
      tags: [quotes]
      summary: Create an immutable executable Quote
      description: Requires `quotes:read`. Resolves requested policy to one concrete mode.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateQuoteRequest'}
      responses:
        '201':
          description: Quote
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Quote'}
        default: {$ref: '#/components/responses/Error'}
  /v1/quotes/{id}:
    parameters:
      - {$ref: '#/components/parameters/QuoteID'}
    get:
      tags: [quotes]
      summary: Read a Quote
      responses:
        '200':
          description: Quote
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Quote'}
        default: {$ref: '#/components/responses/Error'}
  /v1/invocations:
    post:
      tags: [invocations]
      summary: Invoke a bounded Capability
      description: Requires `invocations:create`. Trust mode is inherited from the Quote.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SubmitWorkRequest'}
      responses:
        '200':
          description: Completed, accepted or input-required result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubmitWorkResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/invocations/{id}:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
    get:
      tags: [invocations]
      summary: Get an Invocation-backed Job
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Job'}
        default: {$ref: '#/components/responses/Error'}
  /v1/jobs:
    post:
      tags: [jobs]
      summary: Create asynchronous or interactive work
      description: Requires `jobs:create`. Trust mode is inherited from the Quote.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SubmitWorkRequest'}
      responses:
        '200':
          description: Job accepted or input required
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SubmitWorkResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
    get:
      tags: [jobs]
      summary: Get Job state and proof progress
      description: Requires `jobs:read` and ownership.
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Job'}
        default: {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}/stream:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
      - name: next_sequence
        in: query
        description: Resume from this durable journal sequence (0 = from the start, in which case next_offset and expected_stream_digest must be omitted).
        schema: {type: integer, minimum: 0, default: 0}
      - name: next_offset
        in: query
        description: Required whenever next_sequence > 0 -- must exactly equal the durable cumulative output byte offset as of next_sequence, which is legitimately 0 if no OutputChunk event has occurred yet (e.g. resuming right after an initial STATE event).
        schema: {type: integer, minimum: 0, default: 0}
      - name: expected_stream_digest
        in: query
        description: Required whenever next_sequence > 0 -- must exactly equal the durable cumulative output digest as of next_sequence, which is legitimately empty if no OutputChunk event has occurred yet (e.g. resuming right after an initial STATE event).
        schema: {type: string}
    get:
      tags: [jobs]
      summary: Stream a Job's durable, resumable event journal
      description: Requires `jobs:read` and ownership. Server-Sent Events; each `data:` line is a JobStreamEvent. Delivery is bounded -- reads are paginated server-side rather than loading the full journal at once.
      responses:
        '200':
          description: text/event-stream of JobStreamEvent objects, one per SSE `data:` line
          content:
            text/event-stream:
              schema: {$ref: '#/components/schemas/JobStreamEvent'}
        default: {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}/cancel:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
    post:
      tags: [jobs]
      summary: Cancel a non-terminal Job
      description: Requires `jobs:cancel`. Cancellation preserves the Quote's settlement guarantees.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [idempotency_key]
              properties:
                reason: {type: string}
                idempotency_key: {type: string, minLength: 1}
      responses:
        '200':
          description: Canceled Job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Job'}
        default: {$ref: '#/components/responses/Error'}
  /v1/confirmations/{code}:
    parameters:
      - {name: code, in: path, required: true, schema: {type: string}}
    get:
      tags: [confirmations]
      summary: Read a spend confirmation owned by the caller
      responses:
        '200': {description: Bound spend confirmation}
        default: {$ref: '#/components/responses/Error'}
  /v1/confirmations/{code}/decision:
    parameters:
      - {name: code, in: path, required: true, schema: {type: string}}
    post:
      tags: [confirmations]
      summary: Approve or deny an exact paid request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision]
              properties:
                decision: {type: string, enum: [approve, deny]}
      responses:
        '200': {description: Decision recorded}
        default: {$ref: '#/components/responses/Error'}
  /v1/account:
    get:
      tags: [account]
      summary: Read balance, spend policy and trust policy
      description: Requires `account:read`.
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        default: {$ref: '#/components/responses/Error'}
  /v1/account/usage:
    get:
      tags: [account]
      summary: Read aggregate usage
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        default: {$ref: '#/components/responses/Error'}
  /v1/account/receipts:
    get:
      tags: [receipts]
      summary: List the caller's receipts
      responses:
        '200':
          description: Receipts
          content:
            application/json:
              schema:
                type: object
                required: [receipts]
                properties:
                  receipts: {type: array, items: {$ref: '#/components/schemas/Receipt'}}
        default: {$ref: '#/components/responses/Error'}
  /v1/receipts/{id}:
    parameters:
      - {$ref: '#/components/parameters/ReceiptID'}
    get:
      tags: [receipts]
      summary: Read one owned Receipt
      responses:
        '200':
          description: Receipt
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Receipt'}
        default: {$ref: '#/components/responses/Error'}
  /v1/receipts/{id}/settlement-proof:
    parameters:
      - {$ref: '#/components/parameters/ReceiptID'}
    get:
      tags: [receipts]
      summary: Materialize advanced proof information
      description: Requires `proofs:read`. Managed mock evidence is explicitly not represented as a TOS proof.
      responses:
        '200':
          description: Proof package or Managed evidence descriptor
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        default: {$ref: '#/components/responses/Error'}
  /v1/jobs/{id}/billing:
    parameters:
      - {$ref: '#/components/parameters/JobID'}
    get:
      tags: [earnings]
      summary: Read the durable metered billing snapshot for one Job
      description: >-
        Requires `jobs:read`. Visible to the Job's owning principal (the
        payer) or its owning provider only. Deterministic function of the
        frozen Quote terms and the verified Execution Receipt usage --
        `0 <= gross_charge <= quote.price.total_max` always.
      responses:
        '200':
          description: Billing snapshot
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BillingSnapshot'}
        default: {$ref: '#/components/responses/Error'}
  /v1/provider/earnings:
    get:
      tags: [earnings]
      summary: List the authenticated provider's earnings
      description: Requires `earnings:read`. Always scoped to the caller's own provider identity.
      responses:
        '200':
          description: Earnings
          content:
            application/json:
              schema:
                type: object
                required: [earnings]
                properties:
                  earnings: {type: array, items: {$ref: '#/components/schemas/ProviderEarning'}}
        default: {$ref: '#/components/responses/Error'}
  /v1/provider/earnings/{id}:
    parameters:
      - {$ref: '#/components/parameters/EarningID'}
    get:
      tags: [earnings]
      summary: Read one owned earning
      description: Requires `earnings:read`. 403 if the earning belongs to a different provider.
      responses:
        '200':
          description: Provider earning
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProviderEarning'}
        default: {$ref: '#/components/responses/Error'}
  /v1/uploads:
    post:
      tags: [artifacts]
      summary: Request a signed upload target
      description: '`job_input` requires `invocations:create` or `jobs:create`; `capability_asset` requires `capabilities:write`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUploadRequest'}
      responses:
        '201':
          description: Signed upload target
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UploadTarget'}
        default: {$ref: '#/components/responses/Error'}
  /v1/uploads/{id}/complete:
    parameters:
      - {$ref: '#/components/parameters/ArtifactID'}
    post:
      tags: [artifacts]
      summary: Complete an owned upload
      responses:
        '200':
          description: Available Artifact metadata and content commitment
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StoredArtifact'}
        default: {$ref: '#/components/responses/Error'}
  /v1/artifacts/{id}:
    parameters:
      - {$ref: '#/components/parameters/ArtifactID'}
    get:
      tags: [artifacts]
      summary: Read authorized Artifact metadata
      responses:
        '200':
          description: Artifact metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StoredArtifact'}
        default: {$ref: '#/components/responses/Error'}
  /v1/artifacts/{id}/download-url:
    parameters:
      - {$ref: '#/components/parameters/ArtifactID'}
    get:
      tags: [artifacts]
      summary: Issue a short-lived authorized download URL
      responses:
        '200':
          description: Download target
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DownloadTarget'}
        default: {$ref: '#/components/responses/Error'}
  /v1/taxonomy:
    get:
      tags: [metadata]
      summary: List Capability tags
      responses:
        '200':
          description: Taxonomy
          content:
            application/json:
              schema:
                type: object
                required: [tags]
                properties:
                  tags: {type: array, items: {type: string}}
        default: {$ref: '#/components/responses/Error'}
  /v1/network/status:
    get:
      tags: [metadata]
      summary: Read high-level trust-mode availability
      responses:
        '200':
          description: Network and mode status
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        default: {$ref: '#/components/responses/Error'}
  /v1/providers/{id}/agent-card:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get:
      tags: [metadata]
      summary: Read an ATOS provider Agent Card
      responses:
        '200':
          description: Provider Agent Card
          content:
            application/json:
              schema: {type: object, additionalProperties: true}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque ATOS access token
...
  responses:
    Error:
      description: ATOS error envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    RequestedTrustMode:
      type: string
      enum: [managed, verified, native, auto]
      default: auto
...
    DeviceAuthorizationRequest:
      type: object
      additionalProperties: false
      properties:
        client_type: {type: string}
        client_name: {type: string}
        requested_scopes: {type: array, uniqueItems: true, items: {type: string}}
    DeviceAuthorizationResponse:
      type: object
      required: [device_code, user_code, verification_uri, verification_uri_complete, expires_in, interval]
      properties:
        device_code: {type: string}
        user_code: {type: string}
        verification_uri: {type: string, format: uri}
        verification_uri_complete: {type: string, format: uri}
        expires_in: {type: integer}
        interval: {type: integer}
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token, principal_id, device_id, scopes]
      properties:
        access_token: {type: string}
        token_type: {type: string, const: Bearer}
        expires_in: {type: integer}
        refresh_token: {type: string}
        principal_id: {type: string}
        device_id: {type: string}
        scopes: {type: array, items: {type: string}}
    OAuthError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {type: string, enum: [invalid_request, invalid_grant, authorization_pending, slow_down, access_denied, expired_token, temporarily_unavailable]}
        error_description: {type: string}
...